ti/requirements.txt
# python3

django
gunicorn
gevent
django-cors-headers
requests
django-extensions
djangorestframework
django-filter
# drf-extensions
# pymysql
pyjwt
djangorestframework-jwt
# redis
# celery
# flower
eave

